Gurobi
Co-authored-by: alanctprado <alanctprado@gmail.com>

Initial Gurobi formulation

Gurobi compilation

Initial Gurobi solver

The program is compiling and working. Changes made in the logic of the
base solver (unique_ptr for the ip solver superclass, which is
initialized with one of each subclasses). Required the implementation of
a base `IntegerProgrammingSolver` class. Minor changes to the `options`.

Silenced Gurobi

[WIP] Help me, personal friend 🙏

Come to our side, we have OR-TOOLS

and 🍪

[WIP] Houston, we have an experiment upon us 🤩

Update eval_set.sh

[WIP] Trying to make this work